Complete exercises 01, 06-11
Four new exercises were added to the upstream repository, they are
exercises 02-05. So the old exercises 02-07 were changed to be
exercises 06-11.

This commit simply recognizes this change in exercise ordering. As far
as I can remember, the new exercises were originally completed by me on
Replit, so I will not be completing them again here.